Extended permission control when fonds are created/updated over WS

+++ elza/elza-core/src/main/java/cz/tacr/elza/service/ArrangementService.java
@@ -426,7 +426,10 @@ public class ArrangementService {
     /**
      * @param fund
      * @param ruleSet
-     * @param scopes  
+     * @param scopes
+     * @param userIds
+     * @param groupIds
+     * @param adminPermissionMode strategy for synchronizing the supplied admin users/groups
      * @return Upravená archivní pomůcka
      */
     @Transactional
@@ -435,7 +438,8 @@ public class ArrangementService {
 			                          final RulRuleSet ruleSet,
 			                          final List<ApScope> scopes,
 			                          final List<Integer> userIds,
-			                          final List<Integer> groupIds) {
+			                          final List<Integer> groupIds,
+			                          final AdminPermissionUpdateMode adminPermissionMode) {
         Validate.notNull(fund, "AS musí být vyplněn");
         Validate.notNull(ruleSet, "Pravidla musí být vyplněna");
 
@@ -475,11 +479,11 @@ public class ArrangementService {
         }
 
         if (userIds != null) {
-            syncUsers(originalFund, userIds);
+            syncUsers(originalFund, userIds, adminPermissionMode);
         }
 
         if (groupIds != null) {
-            syncGroups(originalFund, groupIds);
+            syncGroups(originalFund, groupIds, adminPermissionMode);
         }
 
         eventNotificationService
@@ -534,8 +538,10 @@ public class ArrangementService {
      *
      * @param fund
      * @param userIds
+     * @param adminPermissionMode whether to remove existing users not in the supplied list
      */
-    private void syncUsers(final ArrFund fund, final Collection<Integer> userIds) {
+    private void syncUsers(final ArrFund fund, final Collection<Integer> userIds,
+                           final AdminPermissionUpdateMode adminPermissionMode) {
         Validate.notNull(fund, "AS musí být vyplněn");
 
         List<UsrUser> users = userRepository.findByFund(fund);
@@ -551,7 +557,9 @@ public class ArrangementService {
             }
         }
 
-        usersById.values().forEach(u -> userService.deleteUserFundPermissions(u, fund.getFundId()));
+        if (adminPermissionMode == AdminPermissionUpdateMode.FULL_SYNC) {
+            usersById.values().forEach(u -> userService.deleteUserFundPermissions(u, fund.getFundId()));
+        }
     }
 
     /**
@@ -560,8 +568,10 @@ public class ArrangementService {
      *
      * @param fund
      * @param groupIds
+     * @param adminPermissionMode whether to remove existing groups not in the supplied list
      */
-    private void syncGroups(final ArrFund fund, final Collection<Integer> groupIds) {
+    private void syncGroups(final ArrFund fund, final Collection<Integer> groupIds,
+                            final AdminPermissionUpdateMode adminPermissionMode) {
         Validate.notNull(fund, "AS musí být vyplněn");
 
         List<UsrGroup> groups = groupRepository.findByFund(fund);
@@ -577,7 +587,9 @@ public class ArrangementService {
             }
         }
 
-        groupsById.values().forEach(g -> userService.deleteGroupFundPermissions(g, fund.getFundId()));
+        if (adminPermissionMode == AdminPermissionUpdateMode.FULL_SYNC) {
+            groupsById.values().forEach(g -> userService.deleteGroupFundPermissions(g, fund.getFundId()));
+        }
     }
 
     /**
